requirements.txt
dj-database-url-0.5.0 
django-heroku-0.3.1 
whitenoise-5.1.0
africastalking==1.1.8
amqp==2.5.0
Beaker==1.11.0
beautifulsoup4==4.9.1
billiard==3.6.0.0
boto3==1.9.209
botocore==1.12.209
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
cliff==3.1.0
cmd2==0.8.9
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==2.2.2
django-admin-daterange-filter==0.1.3
django-admin-rangefilter==0.5.2
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-envie==1.0.0
django-extensions==2.1.9
django-filter==2.1.0
django-postgrespool2==1.0.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
docutils==0.15.2
emoji==0.5.4
facebook-sdk==3.1.0
google==2.0.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
logutils==0.3.5
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
mypy==0.740
mypy-extensions==0.4.3
openapi-codec==1.3.2
pbr==5.4.5
pecan==1.3.3
Pillow==6.1.0
prettytable==0.7.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycryptodome==3.8.2
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyperclip==1.8.0
PySocks==1.7.0
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.2.1
python-memcached==1.59
pytz==2019.1
PyYAML==5.1.1
raven==6.10.0
redis==3.2.1
reportlab==3.5.34
requests==2.22.0
s3transfer==0.2.1
schema==0.7.1
sendgrid==6.1.0
sentry-sdk==0.10.2
simplejson==3.16.0
six==1.12.0
soupsieve==2.0.1
SQLAlchemy==1.3.11
sqlparse==0.3.0
stevedore==2.0.0
twilio==6.29.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
waitress==1.4.4
wcwidth==0.2.3
WebOb==1.8.6
WebTest==2.0.35
